+++ HarambeeCommerceApi/WeatherForecast.cs
@@ -4,9 +4,9 @@ public class WeatherForecast
 {
     public DateTime Date { get; set; }
 
-    public int TemperatureC { get; set; }
+    public long TemperatureC { get; set; }
 
-    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
+    public long TemperatureF => 32 + (long)(TemperatureC / 0.5556);
 
     public string? Summary { get; set; }
 }
